Fix clippy warnings to reduce noise in build output
This commit addresses multiple categories of clippy warnings to improve code quality and reduce noise:

- Replaced manual div_ceil implementations with standard library method (4 instances)
- Fixed unnecessary if let patterns by using flatten() and iterator methods
- Fixed loop indexing warnings by using iterators instead of indexing
- Removed unnecessary type casts (usize -> usize, u64 -> u64)
- Fixed redundant closures by using direct function references (6 instances)
- Prefixed unused test variables with underscore
- Changed Option.and_then to map when closure always returns Some
- Changed ok_or_else to ok_or when closure doesn't capture variables

These changes reduce library warnings from 57 to approximately 40, making the build output cleaner while maintaining the same functionality.

+++ src/filesystem_service/buffered_file_handle.rs
@@ -281,7 +281,7 @@ impl BufferedFileHandleInner {
         let mut partial = 0;
         let mut total = 0;
 
-        let max_stripe_size = self.config.max_stripe_size as usize;
+        let max_stripe_size = self.config.max_stripe_size;
 
         for builder in self.builders.values() {
             let size = builder.size();
@@ -1074,7 +1074,7 @@ impl BufferedFileHandle {
             let (complete, partial): (Vec<_>, Vec<_>) = inner
                 .builders
                 .iter()
-                .partition(|(_, builder)| builder.size() == max_stripe_size as usize);
+                .partition(|(_, builder)| builder.size() == max_stripe_size);
 
             trace!(
                 force = %force,
